requirements.txt
amqp==2.4.1
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11.12
django-crispy-forms==1.7.2
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-ranged-response==0.2.0
django-redis==4.10.0
django-reversion==3.0.2
django-simple-captcha==0.5.10
djangoajax==3.0.3
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.24.1
future==0.17.1
get==2018.11.19
greenlet==0.4.15
httplib2==0.9.2
idna==2.8
jdcal==1.4
jieba==0.39
kombu==4.3.0
monotonic==1.5
odfpy==1.4.0
openpyxl==2.5.14
Pillow==5.4.1
post==2018.11.20
public==2018.11.20
pycryptodomex==3.7.2
PyMySQL==0.9.3
pysolr==3.8.1
python-alipay-sdk==1.9.1
pytz==2018.9
PyYAML==3.13
query-string==2018.11.20
redis==2.10.6
requests==2.21.0
six==1.12.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24.1
vine==1.2.0
Whoosh==2.7.4
#xadmin==0.6.1
xlrd==1.2.0
xlwt==1.3.0
